% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_names.R
\docType{data}
\name{cpt_names}
\alias{cpt_names}
\title{Names of all colour gradients}
\format{
A character vector.
}
\source{
\url{https://phillips.shef.ac.uk/pub/cpt-city/}
}
\usage{
data(cpt_names)
}
\description{
A character vector containing the names of every colour gradient
bundled in the \pkg{cptcity} package. Use \code{\link{find_cpt}} to
search this vector by keyword.
}
\examples{
data(cpt_names)
head(cpt_names)
length(cpt_names)
}
\keyword{datasets}
