% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_names.R
\docType{data}
\name{cpt_names}
\alias{cpt_names}
\title{Names of the 7140 color gradients of cptcity R Package}
\format{
A vector with the 7140 names of the color gradients
}
\source{
\url{http://soliton.vm.bytemark.co.uk/pub/cpt-city/}
}
\usage{
data(cpt_names)
}
\description{
This dataset os a vector with al the names of the gradients of the archive
cptcity (http://soliton.vm.bytemark.co.uk/pub/cpt-city/) availale in this
package. Please, read the documentation of each color gradient in the
web page shown above.
}
\keyword{datasets}
