% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptcity.R
\docType{package}
\name{cptcity}
\alias{cptcity}
\alias{cptcity-package}
\title{cpt-city: Colour Gradients for R}
\description{
Incorporates 7140 colour gradients from the
\href{http://soliton.vm.bytemark.co.uk/pub/cpt-city/}{cpt-city web archive}
into R. Each palette is stored as a \code{data.frame} with columns
\code{id}, \code{r}, \code{g}, \code{b}.
}
\section{Functions}{

\describe{
  \item{\code{\link{cpt}}}{Return an interpolated colour palette or
    \code{colorRampPalette} function.}
  \item{\code{\link{find_cpt}}}{Search palette names by keyword.}
  \item{\code{\link{lucky}}}{Pick a random palette —
    "I'm Feeling Lucky" for colours.}
  \item{\code{\link{show_cpt}}}{Display multiple palettes side-by-side.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ibarraespinosa/cptcity}
  \item Report bugs at \url{https://github.com/ibarraespinosa/cptcity/issues/}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra-espinosa@noaa.gov} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Authors:
\itemize{
  \item Sergio Ibarra-Espinosa \email{sergio.ibarra-espinosa@noaa.gov} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})
}

}
