% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptcity.R
\name{cptcity}
\alias{cptcity}
\title{Function to return colour palletes from cpt-city=}
\usage{
cptcity(pal = "inferno")
}
\arguments{
\item{pal}{Pallete of colors available or the number of the position
of the same}
}
\value{
A colour pallete.
}
\description{
This function return a collor pallete based on the name or position of the
pallete. Currently there are three available palletes:
}
\details{
\enumerate{
  \item "inferno"
  \item "temperature"
  \item "sst"
  \item "temp"
  \item "arendarl_arctic"
  \item "arendarl_temperature"
  \item "arendarl_zambezi_focus"
  \item "arendarl_zambezi_proximity"
}

The palletes are ones of the most popular.

The palletes are available here:
http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html
}
\examples{
\dontrun{
# Do not run
library(cptcity)
inferno <- cptcity("inferno")
image(matrix(1:255), col = inferno)
}
}
