% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptcity.R
\docType{package}
\name{cptcity}
\alias{cptcity}
\title{A package to return colour gradients from CPTCITY}
\description{
Colour palettes comes from http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html
Rhw function \code{\link{cpt}} has two arguments \strong{n} for the numbers
and \strong{pal} for the name or number of the palette:
}
\details{
The palettes are available here:
http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html
}
