% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cpt.R
\name{show_cpt}
\alias{show_cpt}
\title{Show me some colors!}
\usage{
show_cpt(x, label = TRUE)
}
\arguments{
\item{x}{character;  names of the cpt gradients.}

\item{label}{Logical, to show labels or not.}
}
\value{
names color matrix
}
\description{
\code{\link{show_cpt}} returns a color matrix which prints
and show the colors
}
\examples{
{
library(cptcity)
show_cpt(find_cpt("radar"))
}
}
